Remove cache mechanism on Unifile side
