+++ app/src/main/kotlin/com/hippo/unifile/RawFile.kt
@@ -22,10 +22,10 @@ import android.os.ParcelFileDescriptor.parseMode
 import android.webkit.MimeTypeMap
 import java.io.File
 
-class RawFile(parent: RawFile?, private val file: File) : CachingFile<RawFile>(parent) {
-    override fun list() = file.listFiles()?.let { children ->
+class RawFile(parent: RawFile?, private val file: File) : FileNode<RawFile>(parent) {
+    override fun listFiles() = file.listFiles()?.let { children ->
         MutableList(children.size) { RawFile(this, children[it]) }
-    }
+    } ?: emptyList()
 
     private fun createFile(displayName: String, isFile: Boolean): UniFile? {
         val child = findFile(displayName)
@@ -36,7 +36,7 @@ class RawFile(parent: RawFile?, private val file: File) : CachingFile<RawFile>(p
             val target = File(file, displayName)
             val created = if (isFile) target.createNewFile() else target.mkdir()
             if (created) {
-                RawFile(this, target).also { popCacheIfPresent(it) }
+                RawFile(this, target)
             } else {
                 null
             }
@@ -71,11 +71,9 @@ class RawFile(parent: RawFile?, private val file: File) : CachingFile<RawFile>(p
 
     override fun canWrite() = file.canWrite()
 
-    override fun resolve(displayName: String) = (if (cachePresent) findFile(displayName) else null) ?: RawFile(this, File(file, displayName))
+    override fun resolve(displayName: String) = RawFile(this, File(file, displayName))
 
-    override fun delete() = file.deleteRecursively().also {
-        if (it) parent?.evictCacheIfPresent(name)
-    }
+    override fun delete() = file.deleteRecursively()
 
     override fun exists() = file.exists()
 
@@ -83,10 +81,7 @@ class RawFile(parent: RawFile?, private val file: File) : CachingFile<RawFile>(p
         val old = name
         val target = File(file.parentFile, displayName)
         return if (file.renameTo(target)) {
-            RawFile(parent, target).also { new ->
-                parent?.evictCacheIfPresent(old)
-                parent?.popCacheIfPresent(new)
-            }
+            RawFile(parent, target)
         } else {
             null
         }
